+++ docs/misc/changelog.rst
@@ -4,7 +4,7 @@ Changelog
 ==========
 
 
-Release 1.8.0a2 (WIP)
+Release 1.8.0a3 (WIP)
 --------------------------
 
 
@@ -14,6 +14,8 @@ Breaking Changes:
 
 New Features:
 ^^^^^^^^^^^^^
+- Added ``repeat_action_probability`` argument in ``AtariWrapper``.
+- Only use ``NoopResetEnv`` and ``MaxAndSkipEnv`` when needed in ``AtariWrapper``
 
 `SB3-Contrib`_
 ^^^^^^^^^^^^^^
@@ -23,6 +25,7 @@ New Features:
 
 Bug Fixes:
 ^^^^^^^^^^
+- Fixed Atari wrapper that missed the reset condition (@luizapozzobon)
 
 Deprecations:
 ^^^^^^^^^^^^^
@@ -36,6 +39,7 @@ Others:
 Documentation:
 ^^^^^^^^^^^^^^
 - Renamed ``load_parameters`` to ``set_parameters`` (@DavyMorgan)
+- Fixed typo in ``A2C`` docstring (@AlexPasqua)
 
 
 Release 1.7.0 (2023-01-10)
@@ -1218,4 +1222,4 @@ And all the contributors:
 @Gregwar @ycheng517 @quantitative-technologies @bcollazo @git-thor @TibiGG @cool-RR @MWeltevrede
 @Melanol @qgallouedec @francescoluciano @jlp-ue @burakdmb @timothe-chaumont @honglu2875 @yuanmingqi
 @anand-bala @hughperkins @sidney-tio @AlexPasqua @dominicgkerr @Akhilez @Rocamonde @tobirohrer @ZikangXiong
-@DavyMorgan
+@DavyMorgan @luizapozzobon